+++ bootstrap3/forms.py
@@ -2,14 +2,18 @@
 from __future__ import unicode_literals
 
 from django.contrib.admin.widgets import AdminFileWidget
-from django.forms import HiddenInput, FileInput, CheckboxSelectMultiple, Textarea, TextInput, DateInput, Select
-from django.forms.formsets import BaseFormSet
+from django.forms import (
+    HiddenInput, FileInput, CheckboxSelectMultiple, Textarea, TextInput
+)
 
-from .bootstrap import get_bootstrap_setting, get_form_renderer, get_field_renderer
+from .bootstrap import (
+    get_bootstrap_setting, get_form_renderer, get_field_renderer,
+    get_formset_renderer
+)
 from .text import text_concat, text_value
 from .exceptions import BootstrapError
 from .html import add_css_class, render_tag
-from .icons import render_icon
+from .components import render_icon
 
 
 FORM_GROUP_CLASS = 'form-group'
@@ -19,26 +23,40 @@ def render_formset(formset, **kwargs):
     """
     Render a formset to a Bootstrap layout
     """
-    if not isinstance(formset, BaseFormSet):
-        raise BootstrapError('Parameter "formset" should contain a valid Django FormSet.')
-    forms = [render_form(f, **kwargs) for f in formset]
-    return text_value(formset.management_form) + '\n' + '\n'.join(forms)
+    renderer_cls = get_formset_renderer(**kwargs)
+    return renderer_cls(formset, **kwargs).render()
 
 
-def render_form(form, layout='', **kwargs):
+def render_formset_errors(form, **kwargs):
+    """
+    Render formset errors to a Bootstrap layout
+    """
+    renderer_cls = get_formset_renderer(**kwargs)
+    return renderer_cls(form, **kwargs).render_errors()
+
+
+def render_form(form, **kwargs):
     """
     Render a formset to a Bootstrap layout
     """
-    renderer_cls = get_form_renderer(layout)
-    return renderer_cls(form, layout, **kwargs).render()
+    renderer_cls = get_form_renderer(**kwargs)
+    return renderer_cls(form, **kwargs).render()
 
 
-def render_field(field, layout='', **kwargs):
+def render_form_errors(form, type='all', **kwargs):
+    """
+    Render form errors to a Bootstrap layout
+    """
+    renderer_cls = get_form_renderer(**kwargs)
+    return renderer_cls(form, **kwargs).render_errors(type)
+
+
+def render_field(field, **kwargs):
     """
     Render a formset to a Bootstrap layout
     """
-    renderer_cls = get_field_renderer(layout)
-    return renderer_cls(field, layout, **kwargs).render()
+    renderer_cls = get_field_renderer(**kwargs)
+    return renderer_cls(field, **kwargs).render()
 
 
 def render_label(content, label_for=None, label_class=None, label_title=''):
@@ -55,22 +73,50 @@ def render_label(content, label_for=None, label_class=None, label_title=''):
     return render_tag('label', attrs=attrs, content=content)
 
 
-def render_button(content, button_type=None, icon=None):
+def render_button(
+        content, button_type=None, icon=None, button_class='', size='',
+        href=''):
     """
     Render a button with content
     """
-    attrs = {'class': 'btn'}
+    attrs = {}
+    classes = add_css_class('btn', button_class)
+    size = text_value(size).lower().strip()
+    if size == 'xs':
+        classes = add_css_class(classes, 'btn-xs')
+    elif size == 'sm' or size == 'small':
+        classes = add_css_class(classes, 'btn-sm')
+    elif size == 'lg' or size == 'large':
+        classes = add_css_class(classes, 'btn-lg')
+    elif size == 'md' or size == 'medium':
+        pass
+    elif size:
+        raise BootstrapError(
+            'Parameter "size" should be "xs", "sm", "lg" or ' +
+            'empty ("{}" given).'.format(size))
     if button_type:
         if button_type == 'submit':
-            attrs['class'] += ' btn-primary'
-        elif button_type != 'reset' and button_type != 'button':
-            raise BootstrapError('Parameter "button_type" should be "submit", "reset", "button" or empty.')
+            classes = add_css_class(classes, 'btn-primary')
+        elif button_type not in ('reset', 'button', 'link'):
+            raise BootstrapError(
+                'Parameter "button_type" should be "submit", "reset", ' +
+                '"button", "link" or empty  ("{}" given).'.format(button_type))
         attrs['type'] = button_type
+    attrs['class'] = classes
     icon_content = render_icon(icon) if icon else ''
-    return render_tag('button', attrs=attrs, content=text_concat(icon_content, content, separator=' '))
+    if href:
+        attrs['href'] = href
+        tag = 'a'
+    else:
+        tag = 'button'
+    return render_tag(
+        tag, attrs=attrs, content=text_concat(
+            icon_content, content, separator=' '))
 
 
-def render_field_and_label(field, label, field_class='', label_class='', layout='', **kwargs):
+def render_field_and_label(
+        field, label, field_class='', label_for=None, label_class='',
+        layout='', **kwargs):
     """
     Render a field with its label
     """
@@ -84,9 +130,11 @@ def render_field_and_label(field, label, field_class='', label_class='', layout=
         label_class = add_css_class(label_class, 'control-label')
     html = field
     if field_class:
-        html = '<div class="{klass}">{html}</div>'.format(klass=field_class, html=html)
+        html = '<div class="{klass}">{html}</div>'.format(
+            klass=field_class, html=html)
     if label:
-        html = render_label(label, label_class=label_class) + html
+        html = render_label(
+            label, label_for=label_for, label_class=label_class) + html
     return html
 
 
@@ -108,7 +156,10 @@ def is_widget_required_attribute(widget):
         return False
     if not widget.is_required:
         return False
-    if isinstance(widget, (AdminFileWidget, HiddenInput, FileInput, CheckboxSelectMultiple)):
+    if isinstance(
+            widget, (
+                AdminFileWidget, HiddenInput, FileInput,
+                CheckboxSelectMultiple)):
         return False
     return True
 
@@ -120,4 +171,3 @@ def is_widget_with_placeholder(widget):
     These are all derived form TextInput, except for Textarea
     """
     return isinstance(widget, (TextInput, Textarea))
-